sat-math/backend/requirements.txt
fastapi
uvicorn[standard]
sympy
pydantic-settings
python-multipart
scipy

